Update RapidUpload, try upper/lower/random case md5
